# docs/T033_openapi.yaml
openapi: "3.0.3"
info:
  title: GarudaX Platform API
  description: |
    REST API for the AI Powered Commodity Exchange of Mongolia.
    This gateway translates REST/JSON requests to gRPC calls on backend exchange services.
  version: "1.0.0"
  contact:
    name: GarudaX Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.garudax.mn/api/v1
    description: Production
  - url: https://staging-api.garudax.mn/api/v1
    description: Staging
...
security:
  - BearerAuth: []
paths:
...
  /clearing/positions:
    get:
      tags: [Clearing]
      summary: List positions for a participant
      operationId: getPositions
      parameters:
        - name: participant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Position list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Position"
  /clearing/positions/{instrument_id}:
    get:
      tags: [Clearing]
      summary: Get position for specific instrument
      operationId: getPosition
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: participant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Position details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh an expired JWT
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
            request_id:
              type: string
          required: [code, message]
...
    Position:
      type: object
      properties:
        participant_id:
          type: string
          format: uuid
        instrument_id:
          type: string
          format: uuid
        net_quantity:
          type: integer
          description: Positive = long, negative = short
        avg_entry_price:
          type: string
        total_buy_qty:
          type: integer
        total_sell_qty:
          type: integer
        realized_pnl:
          type: string
        updated_at:
          type: string
          format: date-time
...
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: Seconds until access token expires
...